Revert "Network layer"
