--- MarvelApiChallenge/Netwok/NetworkErrors.swift
@@ -1,13 +1,0 @@
-//
-//  NetworkErrors.swift
-//  MarvelApiChallenge
-//
-//  Created by Gabriel Policastro on 24/03/23.
-//
-
-import Foundation
-
-enum NetworkErrors: Error {
-    case generic
-    case decoding
-}
