+++ MarvelApiChallenge/SceneDelegate.swift
@@ -16,8 +16,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(frame: UIScreen.main.bounds)
         self.window?.windowScene = windowScene
-        let viewController = HomeFactory.createViewController()
-        self.window?.rootViewController = viewController
+        let VC = HomeViewController()
+        self.window?.rootViewController = VC
         self.window?.makeKeyAndVisible()
     }
