--- MarvelApiChallenge/Scenes/Home/HomeViewController/HomeFactory.swift
@@ -1,18 +1,0 @@
-//
-//  HomeFactory.swift
-//  MarvelApiChallenge
-//
-//  Created by Gabriel Policastro on 25/03/23.
-//
-
-import UIKit
-
-final class HomeFactory {
-    static func createViewController() -> UIViewController {
-        let provider = NetworkProvider<HomeModels.GetHeroesRequest>()
-        let service = HomeService(networkProvider: provider)
-        let viewModel = HomeViewModel(homeService: service)
-        let viewController = HomeViewController(viewModel: viewModel)
-        return viewController
-    }
-}
