--- MarvelApiChallenge/Utils/Constants.swift
@@ -1,16 +1,0 @@
-//
-//  Constants.swift
-//  MarvelApiChallenge
-//
-//  Created by Gabriel Policastro on 24/03/23.
-//
-
-import Foundation
-
-struct Constants {
-    static let baseURL = "https://gateway.marvel.com:443/v1/public/"
-    static let timeStamp = "1"
-    static let apiKey = "9fbf6495c4d40b6b81df06bf92f12ca1"
-    static let hash = "76ef495656a227983ebed502bab079e9"
-    
-}
