--- MarvelApiChallenge/Utils/Images.swift
@@ -1,14 +1,0 @@
-//
-//  Images.swift
-//  MarvelApiChallenge
-//
-//  Created by Gabriel Policastro on 24/03/23.
-//
-
-import Foundation
-
-struct Images {
-    static let banner = "banner"
-    static let favorite = "favorite"
-}
-
