--- Package.swift
@@ -1,21 +1,0 @@
-// swift-tools-version: 5.7
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
-import PackageDescription
-
-let package = Package(
-    name: "MarvelApiChallenge",
-    dependencies: [
-        .package(url: "https://github.com/Moya/Moya.git", .upToNextMajor(from: "15.0.0"))
-    ],
-    targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages this package depends on.
-        .executableTarget(
-            name: "MarvelApiChallenge",
-            dependencies: ["Moya"]),
-        .testTarget(
-            name: "MarvelApiChallengeTests",
-            dependencies: ["MarvelApiChallenge"]),
-    ]
-)
