--- Sources/MarvelApiChallenge/MarvelApiChallenge.swift
@@ -1,8 +1,0 @@
-@main
-public struct MarvelApiChallenge {
-    public private(set) var text = "Hello, World!"
-
-    public static func main() {
-        print(MarvelApiChallenge().text)
-    }
-}
