--- Tests/MarvelApiChallengeTests/MarvelApiChallengeTests.swift
@@ -1,11 +1,0 @@
-import XCTest
-@testable import MarvelApiChallenge
-
-final class MarvelApiChallengeTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(MarvelApiChallenge().text, "Hello, World!")
-    }
-}
